Moved transitionConfig to own file and made transition instant
