+++ client/app/config/transitions.js
@@ -1,0 +1,27 @@
+import { Animated, Easing } from 'react-native';
+
+const transitionConfig = () => {
+    return {
+      transitionSpec: {
+        duration: 0, // Transition is "instant"
+        easing: Easing.out(Easing.poly(4)),
+        timing: Animated.timing,
+        useNativeDriver: true,
+      },
+      screenInterpolator: sceneProps => {
+        const { layout, position, scene } = sceneProps
+  
+        const thisSceneIndex = scene.index
+        const width = layout.initWidth
+  
+        const translateX = position.interpolate({
+          inputRange: [thisSceneIndex - 1, thisSceneIndex],
+          outputRange: [width, 0],
+        })
+  
+        return { transform: [{ translateX }] }
+      },
+    }
+  }
+
+  export default transitionConfig;
